requirements.txt
langchain
gdown
redis==4.5.3
numpy
openai
pandas
python-dotenv
